pyproject.toml
[project]
name = "creden-mcp"
version = "0.1.0"
description = "MCP server for data.creden.co — Thai company information"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.0",
    "playwright>=1.42.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[project.scripts]
creden-mcp = "creden_mcp.server:main"
creden-discovery = "creden_mcp.discovery:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/creden_mcp"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
